希望地域ファイルの作成、関数の移動
